# Conflicts:
#	Often/MediaItemsViewController.swift

+++ Keyboard/PackMediaItem.swift
@@ -16,7 +16,7 @@ class PackMediaItem: MediaItem {
     var items_count: Int?
     var items: [LyricMediaItem] = []
     var premium: Bool?
-    var price: Int?
+    var price: Double?
     
     required init(data: NSDictionary) {
         super.init(data: data)
@@ -60,7 +60,7 @@ class PackMediaItem: MediaItem {
             self.premium = premium
         }
         
-        if let price = data["price"] as? Int {
+        if let price = data["price"] as? Double {
             self.price = price
         }
     }
